refactor(rolldown_binding): move and split binding_js_or_regex
