+++ crates/rolldown_binding/src/options/binding_input_options/binding_watch_option.rs
@@ -1,4 +1,4 @@
-use crate::options::plugin::types::binding_js_or_regex::{
+use crate::types::binding_string_or_regex::{
   BindingStringOrRegex, bindingify_string_or_regex_array,
 };
