+++ crates/rolldown_binding/src/types/binding_string_or_regex.rs
@@ -1,58 +1,11 @@
 use std::fmt::Debug;
 
-use napi::{
-  Either, Env, Error, JsObject, JsUnknown, NapiValue, Status,
-  bindgen_prelude::{FromNapiValue, Function, TypeName, ValidateNapiValue},
-  sys,
-};
+use napi::{Either, bindgen_prelude::FromNapiValue, sys};
 
 use rolldown_utils::js_regex::HybridRegex;
 use rolldown_utils::pattern_filter::StringOrRegex;
 
-#[derive(Debug, Default, Clone)]
-pub struct JsRegExp {
-  pub source: String,
-  pub flags: String,
-}
-
-impl ValidateNapiValue for JsRegExp {}
-
-impl TypeName for JsRegExp {
-  fn type_name() -> &'static str {
-    "RegExp"
-  }
-
-  fn value_type() -> napi::ValueType {
-    napi::ValueType::Object
-  }
-}
-
-impl FromNapiValue for JsRegExp {
-  unsafe fn from_napi_value(env: sys::napi_env, napi_val: sys::napi_value) -> napi::Result<Self> {
-    let js_object = unsafe { JsObject::from_raw_unchecked(env, napi_val) };
-
-    let env = Env::from(env);
-    let global = env.get_global()?;
-    let regexp_constructor = global.get_named_property::<Function<JsUnknown, ()>>("RegExp")?;
-
-    if js_object.instanceof(regexp_constructor)? {
-      let source = js_object.get_named_property::<String>("source")?;
-      let flags = js_object.get_named_property::<String>("flags")?;
-
-      Ok(JsRegExp { source, flags })
-    } else {
-      Err(Error::new(Status::ObjectExpected, "Expect a RegExp object"))
-    }
-  }
-}
-
-impl TryFrom<JsRegExp> for HybridRegex {
-  type Error = anyhow::Error;
-
-  fn try_from(value: JsRegExp) -> Result<Self, Self::Error> {
-    HybridRegex::with_flags(&value.source, &value.flags)
-  }
-}
+use super::js_regex::JsRegExp;
 
 #[derive(Debug, Clone)]
 pub struct BindingStringOrRegex(StringOrRegex);
